+++ texk/dvipdfm-x/dpxutil.h
@@ -60,7 +60,7 @@ struct ht_entry {
 typedef void (*hval_free_func) (void *);
 
 struct ht_table {
-  long   count;
+  int count;
   hval_free_func hval_free_fn;
   struct ht_entry *table[HASH_TABLE_SIZE];
 };
@@ -68,7 +68,7 @@ struct ht_table {
 extern void  ht_init_table   (struct ht_table *ht,
                               hval_free_func hval_free_fn);
 extern void  ht_clear_table  (struct ht_table *ht);
-extern long  ht_table_size   (struct ht_table *ht);
+extern int   ht_table_size   (struct ht_table *ht);
 extern void *ht_lookup_table (struct ht_table *ht,
                               const void *key, int keylen);
 extern void  ht_append_table (struct ht_table *ht,
